web/requirements.txt
Django==3.0.8
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.0
